Version 1 of working firmware for Volume Up/Down
Not ideal from the power consumption point of view: had to have 2 pins
to be latched to 1. From quick measurements the sleep power consumption
is 14uA. It is above datasheet and above some earlier tests. Maybe there
is a trick to lower it down. As it stands currently, it is about 2 years
in sleep mode. Not too bad, but not totally great either.